# main/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Swagger Petstore - OpenAPI 3.0
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_

    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
- url: http://localhost:8000/
tags:
- name: auth
  description: Access to the system
- name: home
  description: Operations about house
- name: access
  description: Operations about user
- name: devices
  description: Operations about devices
paths:
  /auth/sign-up:
    post:
      tags:
      - auth
      summary: Registration of new users
      description: "Registration of users in the system using a password, email and\
        \ login"
      requestBody:
        description: "Adding new users to the system using a password, email, and\
          \ login"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth_signup_body'
        required: true
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid data value
        "409":
          description: A user with such an email has already been registered
        "500":
          description: Internal error on the server
  /auth/sign-in:
    post:
      tags:
      - auth
      summary: User authentication
      description: User authentication by mail and password
      requestBody:
        description: Search for a user in the system using the specified email and
          password
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth_signin_body'
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
        "400":
          description: Invalid data value
        "401":
          description: Access denied
        "500":
          description: Internal error on the server
  /auth/code:
    post:
      tags:
      - auth
      summary: Update password
      description: Sending the recovery code to the mail
      requestBody:
        description: Sending the recovery code to the specified email address
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth_code_body'
        required: true
      responses:
        "200":
          description: successful operation
        "404":
          description: The resource does not exist
        "500":
          description: Internal error on the server
  /auth/verification:
    post:
      tags:
      - auth
      summary: Checking the sent code
      description: Checking the verification code sent to the mail
      requestBody:
        description: Checking the verification code sent to the mail
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth_verification_body'
        required: true
      responses:
        "200":
          description: successful operation
        "400":
          description: Invalid code
        "401":
          description: Access denied
        "500":
          description: Internal error on the server
  /auth/password:
    put:
      tags:
      - auth
      summary: Update password
      description: Update password
      requestBody:
        description: Checking the entered passwords for a match and updating it in
          the system
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/auth_password_body'
        required: true
      responses:
        "200":
          description: successful operation
        "401":
          description: Access denied
        "500":
          description: Internal error on the server
  /api/homes:
    get:
      tags:
      - home
      summary: List of user's houses
      description: Get a list of the user's houses
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/inline_response_200_1'
                x-content-type: application/json
        "404":
          description: The resource does not exist
        "500":
          description: Internal error on the server
    post:
      tags:
      - home
      summary: Create a new home
      description: Creating a new home by the user
      requestBody:
        description: "Creating a new house by name, geogrpahcoords"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/homes_body'
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_2'
        "400":
          description: Invalid data value
        "401":
          description: Access denied
        "404":
          description: The resource does not exist
        "500":
          description: Internal error on the server
  /api/homes/{homeID}:
    get:
      tags:
      - home
      summary: Get inforamtions about home
      description: Get inforamtions about home by homeid
      parameters:
      - name: homeID
        in: path
        description: ID of the home
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_3'
        "404":
          description: The resource does not exist
        "500":
          description: Internal error on the server
    put:
      tags:
      - home
      summary: Updating the home name
      description: Updating the home name
      parameters:
      - name: homeID
        in: path
        description: ID of the home
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      requestBody:
        description: Updating the home name
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/homes_homeID_body'
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_4'
        "401":
          description: Access denied
        "403":
          description: Access denied due to insufficient permissions
        "404":
          description: The resource does not exist
        "500":
          description: Internal error on the server
    delete:
      tags:
      - home
      summary: Delete home
      description: Removal of the house by the owner by name
      parameters:
      - name: homeID
        in: path
        description: ID of the home
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: successful operation
        "401":
          description: Access denied
        "403":
          description: Access denied due to insufficient permissions
        "404":
          description: The resource does not exist
        "500":
          description: Internal error on the server
  /api/homes/{homeID}/accesses:
    get:
      tags:
      - access
      summary: List of user's houses
      description: Get a list of the user's houses
      parameters:
      - name: homeID
        in: path
        description: ID of the home
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/inline_response_200_5'
                x-content-type: application/json
        "404":
          description: The resource does not exist
        "500":
          description: Internal error on the server
    post:
      tags:
      - access
      summary: Adding a user to the house
      description: Adding a user to the house by specified mail
      parameters:
      - name: homeID
        in: path
        description: ID of the home
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/homeID_accesses_body'
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_6'
        "401":
          description: Access denied
        "403":
          description: Access denied due to insufficient permissions
        "404":
          description: The resource does not exist
        "500":
          description: Internal error on the server
  /api/homes/{homeID}/accesses/{accessID}:
    get:
      tags:
      - access
      summary: Get informations about access
      description: Get informations about access by accessid
      parameters:
      - name: homeID
        in: path
        description: ID of the home
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      - name: accessID
        in: path
        description: ID of the device
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_7'
        "404":
          description: The resource does not exist
        "500":
          description: Internal error on the server
    put:
      tags:
      - access
      summary: Update the user's access level
      description: Update the user's access level
      parameters:
      - name: homeID
        in: path
        description: ID of the home
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      requestBody:
        description: Update the user's access level
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/accesses_accessID_body'
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_8'
        "400":
          description: Invalid data value
        "401":
          description: Access denied
        "403":
          description: Access denied due to insufficient permissions
        "404":
          description: The resource does not exist
        "500":
          description: Internal error on the server
    delete:
      tags:
      - access
      summary: Remove a user from home
      description: Remove a user from home
      parameters:
      - name: homeID
        in: path
        description: ID of the home
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      - name: accessID
        in: path
        description: ID of the device
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: successful operation
        "401":
          description: Access denied
        "403":
          description: Access denied due to insufficient permissions
        "404":
          description: The resource does not exist
        "500":
          description: Internal error on the server
  /api/homes/{homeID}/devices:
    get:
      tags:
      - devices
      summary: List of devices houses
      description: Get a list of the devices houses
      parameters:
      - name: homeID
        in: path
        description: ID of the home
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/inline_response_200_9'
                x-content-type: application/json
        "404":
          description: The resource does not exist
        "500":
          description: Internal error on the server
    post:
      tags:
      - devices
      summary: Add a device
      description: Add a new device to your home
      parameters:
      - name: homeID
        in: path
        description: ID of the home
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      requestBody:
        description: Add a new device to your home
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/homeID_devices_body'
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_10'
        "400":
          description: Invalid data value
        "401":
          description: Access denied
        "403":
          description: Access denied due to insufficient permissions
        "404":
          description: The resource does not exist
        "500":
          description: Internal error on the server
  /api/homes/{homeID}/devices/{deviceID}:
    get:
      tags:
      - devices
      summary: Get informations about device
      description: Get informations about device by deviceid
      parameters:
      - name: homeID
        in: path
        description: ID of the home
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      - name: deviceID
        in: path
        description: ID of the device
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_9'
        "404":
          description: The resource does not exist
        "500":
          description: Internal error on the server
    delete:
      tags:
      - devices
      summary: Remove a device from home
      description: Remove a device from home
      parameters:
      - name: homeID
        in: path
        description: ID of the home
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      - name: deviceID
        in: path
        description: ID of the device
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: successful operation
        "401":
          description: Access denied
        "403":
          description: Access denied due to insufficient permissions
        "404":
          description: The resource does not exist
        "500":
          description: Internal error on the server
  /api/homes/{homeID}/devices/{deviceID}/status:
    put:
      tags:
      - devices
      summary: Start the device
      description: Start the device
      parameters:
      - name: homeID
        in: path
        description: ID of the home
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      - name: deviceID
        in: path
        description: ID of the device
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: successful operation
        "400":
          description: Invalid data value
        "404":
          description: The resource does not exist
        "500":
          description: Internal error on the server
  /api/homes/{homeID}/devices/{deviceID}/history:
    get:
      tags:
      - devices
      summary: List of device history
      description: Get a list of device history
      parameters:
      - name: homeID
        in: path
        description: ID of the home
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      - name: deviceID
        in: path
        description: ID of the device
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: A list of devices' history records.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_11'
        "404":
          description: The resource does not exist
        "500":
          description: Internal error on the server
components:
  schemas:
    auth_signup_body:
      type: object
      properties:
        password:
          type: string
          example: abcd
        name:
          type: string
          example: user
        email:
          type: string
          example: email@mail.ru
    auth_signin_body:
      type: object
      properties:
        password:
          type: string
          example: abcd
        email:
          type: string
          example: email@mail.ru
    inline_response_200:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3MjY3NzgxNjUsImlhdCI6MTcyNjczNDk2NSwidXNlcklkIjoxfQ.aboYTZrft9YR1q59Q7u9AdyVhUDsi5ep4Pfzicde-5Y
        name:
          type: string
          example: user1
        email:
          type: string
          example: email1
      example:
        name: user1
        email: email1
        token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE3MjY3NzgxNjUsImlhdCI6MTcyNjczNDk2NSwidXNlcklkIjoxfQ.aboYTZrft9YR1q59Q7u9AdyVhUDsi5ep4Pfzicde-5Y
    auth_code_body:
      type: object
      properties:
        email:
          type: string
          example: email@mail.ru
    auth_verification_body:
      type: object
      properties:
        code:
          type: string
          example: 1a2b3c
    auth_password_body:
      type: object
      properties:
        password:
          type: string
          example: abcd
    homes_geographcoors:
      type: object
      properties:
        latitude:
          type: number
          format: float
          example: 37.7749
        longitude:
          type: number
          format: float
          example: -122.4194
      example:
        latitude: 37.7749
        longitude: -122.4194
    inline_response_200_1:
      type: object
      properties:
        homeid:
          type: integer
          example: 2
        name:
          type: string
          example: home
        geographcoors:
          $ref: '#/components/schemas/homes_geographcoors'
        accessLevel:
          type: integer
          example: 4
      example:
        accessLevel: 4
        name: home
        homeid: 2
        geographcoors:
          latitude: 37.7749
          longitude: -122.4194
    homes_body:
      type: object
      properties:
        name:
          type: string
          example: home
        geographcoors:
          $ref: '#/components/schemas/homes_geographcoors'
    inline_response_200_2:
      type: object
      properties:
        homeid:
          type: integer
          example: 2
        name:
          type: string
          example: home
        geographcoors:
          type: object
          properties:
            latitude:
              type: number
              format: float
              example: 37.7749
            longitude:
              type: number
              format: float
              example: -122.4194
          example:
            latitude: 37.7749
            longitude: -122.4194
        accessLevel:
          type: integer
          example: 4
      example:
        accessLevel: 4
        name: home
        homeid: 2
        geographcoors:
          latitude: 37.7749
          longitude: -122.4194
    inline_response_200_3:
      properties:
        homeid:
          type: integer
          example: 2
        name:
          type: string
          example: home
        geographcoors:
          type: object
          properties:
            latitude:
              type: number
              format: float
              example: 37.7749
            longitude:
              type: number
              format: float
              example: -122.4194
          example:
            latitude: 37.7749
            longitude: -122.4194
        accessLevel:
          type: integer
          example: 4
      example:
        accessLevel: 4
        name: home
        homeid: 2
        geographcoors:
          latitude: 37.7749
          longitude: -122.4194
    homes_homeID_body:
      type: object
      properties:
        name:
          type: string
          example: home2
    inline_response_200_4:
      type: object
      properties:
        homeid:
          type: integer
          example: 2
        name:
          type: string
          example: home2
        geographcoors:
          type: object
          properties:
            latitude:
              type: number
              format: float
              example: 37.7749
            longitude:
              type: number
              format: float
              example: -122.4194
          example:
            latitude: 37.7749
            longitude: -122.4194
      example:
        name: home2
        homeid: 2
        geographcoors:
          latitude: 37.7749
          longitude: -122.4194
    inline_response_200_5:
      type: object
      properties:
        accessod:
          type: integer
          example: 2
        name:
          type: string
          example: user1
        email:
          type: string
          example: user1@example.com
        accessLevel:
          type: string
          example: owner
      example:
        accessLevel: owner
        accessod: 2
        name: user1
        email: user1@example.com
    homeID_accesses_body:
      type: object
      properties:
        email:
          type: string
          example: 123@mail.ru
        accessLevel:
          type: integer
          example: 2
    inline_response_200_6:
      properties:
        accessid:
          type: integer
          example: 2
        name:
          type: string
          example: user2
        email:
          type: string
          example: email2
        accesslevel:
          type: integer
          example: 4
      example:
        accessid: 2
        accesslevel: 4
        name: user2
        email: email2
    inline_response_200_7:
      type: object
      properties:
        accessid:
          type: integer
          example: 3
        name:
          type: string
          example: user1
        email:
          type: string
          example: user1@example.com
        accessLevel:
          type: integer
          example: 4
      example:
        accessid: 3
        accessLevel: 4
        name: user1
        email: user1@example.com
    accesses_accessID_body:
      type: object
      properties:
        email:
          type: string
          example: email1
        level:
          type: integer
          example: 3
    inline_response_200_8:
      type: object
      properties:
        accessid:
          type: integer
          example: 3
        name:
          type: string
          example: user1
        email:
          type: string
          example: email1
        accessLevel:
          type: integer
          example: 4
      example:
        accessid: 3
        accessLevel: 4
        name: user1
        email: email1
    inline_response_200_9:
      type: object
      properties:
        deviceid:
          type: integer
          example: 2
        name:
          type: string
          example: dev1
        brand:
          type: string
          example: BrandX
        status:
          type: string
          example: active
      example:
        name: dev1
        deviceid: 2
        brand: BrandX
        status: active
    homeID_devices_body:
      type: object
      properties:
        name:
          type: string
          example: dev1
    inline_response_200_10:
      type: object
      properties:
        deviceid:
          type: integer
          example: 2
        name:
          type: string
          example: dev1
        brand:
          type: string
          example: BrandX
        status:
          type: string
          example: inactive
      example:
        name: dev1
        deviceid: 2
        brand: BrandX
        status: inactive
    inline_response_200_11:
      type: object
      properties:
        name:
          type: string
          example: dev1
        history:
          type: array
          items:
            type: object
            properties:
              timework:
                type: integer
                description: Total time worked by the device
                example: 120
              average:
                type: number
                description: Average indicator value
                format: float
                example: 23.5
              energy:
                type: integer
                description: Energy consumed by the device
                example: 300
            example:
              average: 23.5
              timework: 120
              energy: 300
      example:
        name: dev1
        history:
        - average: 23.5
          timework: 120
          energy: 300
        - average: 23.5
          timework: 120
          energy: 300